feat: add robot

+++ git/setup.sh
@@ -14,13 +14,22 @@ if [[ ! -d $path_git_template_hook ]]; then
 fi
 
 # 确保 git-templates 的配置
-git config --global init.templatedir $path_git_template
-# 将 post-checkout 内容拷贝过去
-cp "${path_script_self}/post-checkout" $path_git_template_hook
+if [[ -z $(git config --global init.templatedir) ]]; then
+  git config --global init.templatedir $path_git_template
+fi
+
+# githook post-checkout setup
+if [[ ! -x "${path_git_template_hook}/post-checkout" ]]; then
+  # 将 post-checkout 内容拷贝过去
+  cp "${path_script_self}/post-checkout" $path_git_template_hook
+  # 加权，保证可执行
+  chmod +x "$path_git_template_hook/post-checkout"
+fi
+
 # 默认 git identity
-cp "${path_script_self}/git-identity" "${HOME}/.git-identity"
-# 加权，保证可执行
-chmod +x "$path_git_template_hook/post-checkout"
+if [[ ! -f "${HOME}/.git-identity" ]]; then
+  cp "${path_script_self}/git-identity" "${HOME}/.git-identity"
+fi
 
 # say goodbye
 echo "git-templates 与 hooks 已配置完毕"
